Restructure project, added php, db, css, img, and js folders. Changed index.php to index.html and content is now a mostly finished landing page.
