--- another_test.txt
@@ -1,1 +1,0 @@
-another one
