--- classified_item.php
@@ -1,27 +1,0 @@
-<?php
-require_once "SLCCGAuthenticate.php";
-?>
-<!DOCTYPE html>
-<html>
-<head>
-<title>SLCC Gardens - </title>
-<style type="text/css">
-nav * {vertical-align:middle;}
-</style>
-<meta charset="utf-8"/>
-<meta name="viewport" content="width=device-width, initial-scale=1.0"> 
-<script src="js/jquery-3.3.1.min.js"></script>
-<script src="js/jquery-ui.min.js"></script>
-</head>
-<body>
-<nav>
-  <img src="images/static/slccg.png" />
-  <a href="classifieds_landing.php">Classifieds</a> 
-  | <a href="gardenbox_landing.php">Garden Box Logs</a> 
-  | <a href="manage_members.php">Manage Members</a> 
-  | <a href="index.php">Back to Main Site</a> 
-</nav>
-<h1>This page shows details on a classified item.</h1>
-<img src="images/classifieds/pumpkins.jpg" />
-</body>
-</html>
