--- index.php
@@ -1,60 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-<title>SLCC Gardens - Members Pages</title>
-<style type="text/css">
-html {font-size:16px; font-family:Helvetica, Arial, sans-serif;}
-div.about {border-radius:1em; margin:1em; padding:1em; 
-  border:2px solid #000000; max-width:40em;}
-div.login {float:right; margin:1em; padding:1em; border-radius:1em;
-  border:1px solid #000000; background-color:#ff3333;}
-div.login a {margin:1em; padding:0.5em; color:#000000; 
-  text-decoration:none; border-radius:0.5em;}
-div.login:hover a {background-color:#ffcccc;}
-</style>
-<meta charset="utf-8"/>
-<meta name="viewport" content="width=device-width, initial-scale=1.0"> 
-<link rel="shortcut icon" type="image/x-icon" href="images/slccgardensfave.ico" />
-<script src="js/jquery-3.3.1.min.js"></script>
-<script src="js/jquery-ui.min.js"></script>
-</head>
-<body>
-
-<h1>Members Page</h1>
-<p>This would be the last page on the Wix site. 
-It describes the Classifieds and the Garden Box pages
-and provides links to them.</p>
-
-<div class="about">
-<div class="login">
-<a href="classifieds_landing.php">Login</a>
-</div>
-<h2>About Classifieds</h2>
-<p><b>A paragraph or two about what classifieds are and how they work.</b>
-Lorem ipsum dolor sit amet, consectetur adipiscing elit. Mauris a 
-est ipsum. Aenean ut lectus id massa pulvinar aliquet. Suspendisse 
-finibus, justo et mollis pharetra, orci risus suscipit nisl, nec 
-consectetur ante velit vitae ante. Cras aliquet massa eget ante 
-pulvinar tempor. Duis gravida, neque quis efficitur iaculis, augue 
-augue accumsan mauris, ac bibendum arcu lorem pellentesque ante. 
-Integer ac ullamcorper dui, sit amet sodales nibh. Donec orci dolor, 
-ornare nec auctor eu, vestibulum vel lorem.</p>
-</div>
-
-<div class="about">
-<div class="login">
-<a href="gardenbox_landing.php">Login</a>
-</div>
-<h2>About Garden Box Pages</h2>
-<p><b>A paragraph or two about what garden box pages are and how they work.</b>
-Lorem ipsum dolor sit amet, consectetur adipiscing elit. Mauris a 
-est ipsum. Aenean ut lectus id massa pulvinar aliquet. Suspendisse 
-finibus, justo et mollis pharetra, orci risus suscipit nisl, nec 
-consectetur ante velit vitae ante. Cras aliquet massa eget ante 
-pulvinar tempor. Duis gravida, neque quis efficitur iaculis, augue 
-augue accumsan mauris, ac bibendum arcu lorem pellentesque ante. 
-Integer ac ullamcorper dui, sit amet sodales nibh. Donec orci dolor, 
-ornare nec auctor eu, vestibulum vel lorem.</p>
-</div>
-</body>
-</html>
