+++ php/Logout.php
@@ -1,0 +1,5 @@
+<?php
+session_start();
+unset($_SESSION);
+session_destroy();
+header('Location:landing.html');
